Finished support for tagged searches.
Currently supports tag searches for low fat, low salt, low sugar, gluten
free or vegetarian recipes.

Tagged searches aren't going to be perfect, mostly because of mistakes
in the underlying recipe database we are using (it seems like they
scraped the web to find their ingredients and weren't perfect with it),
but the majority of recipes should turn out well with tagged searches.