Tags Pt. 2
Potentially completed. Now all tags associated with an account are deleted when the search page is exited (they do not persist like ingredients). This is to avoid a massive headache, and should work fine as long as the "filter" button under the tags allows a user to attempt a new search with the tags they selected active